requirements.txt
# Core Framework
streamlit>=1.28.0

# LangChain
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1

# Embeddings & Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
trafilatura>=1.6.0
readability-lxml>=0.8.1

# Environment & Utilities
python-dotenv>=1.0.0
